First hack at .vimrc file.
Includes all of the features mentioned in the README.md file. All
in one commit because I want a text editor to start with a good
base feature set when I work with it; I don't want to incrementally
add changes and have to enable things like syntax highlighting, and
search highlighting, and so forth, over and over and over. This set
of features is a good start.